refactor (#17)
* - refactor: remove BigInteger from ids type

* - fix: add mapper to MapRouteDataDTO for prepared routes

* - refactor: add 'private' to field where it necessary

* - rafactor: add RouteController and merge NavigationService with RouteService

* + add: custom exception for mapbox

+++ src/main/kotlin/ru/vdnh/repository/EventJdbc.kt
@@ -4,7 +4,6 @@ import org.springframework.jdbc.core.JdbcTemplate
 import org.springframework.stereotype.Repository
 import ru.vdnh.model.entity.EventEntity
 import ru.vdnh.repository.mapper.EventRowMapper
-import java.math.BigInteger
 
 @Repository
 class EventJdbc(
@@ -12,7 +11,7 @@ class EventJdbc(
     private val eventRowMapper: EventRowMapper
 ) : EventRepository {
 
-    override fun findEvent(id: BigInteger): EventEntity {
+    override fun findEvent(id: Long): EventEntity {
         return jdbcTemplate.queryForObject(
             "$SQL_SELECT_ENTITY WHERE e.id = ?", eventRowMapper, id
         )!!
